- Fixing Nodes between planets - Splitting Nodes Between sunNodes & mercuryNodes
